refactor: 가독성을 위해 모듈 분리
