+++ src/game/printResult.js
@@ -1,15 +1,5 @@
-function getRandomNumber() {
-  const min = 0;
-  const max = 10;
-
-  return Math.floor(Math.random() * (max - min)) + min;
-}
-
-function race(car) {
-  if (getRandomNumber() > 3) {
-    car.count++;
-  }
-}
+import race from './race.js';
+import getWinner from './getWinner.js';
 
 function raceResultTemplate(car) {
   return `${car.name}: ${'-'.repeat(car.count)}<br>`;
@@ -19,27 +9,6 @@ function printRaceResult(car) {
   const $resultScreen = document.querySelector('#app > div:last-child');
 
   $resultScreen.insertAdjacentHTML('beforeend', raceResultTemplate(car));
-}
-
-function getMaxCount(carList) {
-  let maxCount = 0;
-  carList.forEach((car) => {
-    if (maxCount < car.count) {
-      maxCount = car.count;
-    }
-  });
-
-  return maxCount;
-}
-
-function getWinner(carList) {
-  const maxCount = getMaxCount(carList);
-  return carList.filter((car) => {
-    if (car.count === maxCount) {
-      return true;
-    }
-    return false;
-  });
 }
 
 function resetResult($resultScreen, carList) {
